# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Autel MaxiSYS MS Ultra S2 Diagnostic Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostics/sessions:
    get:
      operationId: listDiagnosticSessions
      tags: [diagnostics]
      summary: List all diagnostic sessions
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of diagnostic sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagnosticSession"
    post:
      operationId: startDiagnosticSession
      tags: [diagnostics]
      summary: Start a new diagnostic session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartDiagnosticSessionBody"
      responses:
        "201":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostics/sessions/{id}:
    get:
      operationId: getDiagnosticSession
      tags: [diagnostics]
      summary: Get diagnostic session details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Diagnostic session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostics/dtc:
    get:
      operationId: listDtcCodes
      tags: [diagnostics]
      summary: List DTC codes for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DTC codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DtcCode"
  /diagnostics/dtc/{id}/clear:
    post:
      operationId: clearDtcCode
      tags: [diagnostics]
      summary: Clear a specific DTC code
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: DTC cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DtcCode"
...
  /adas/calibrations:
    get:
      operationId: listAdasCalibrations
      tags: [adas]
      summary: List ADAS calibration records
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Calibration records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdasCalibration"
    post:
      operationId: createAdasCalibration
      tags: [adas]
      summary: Start ADAS calibration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdasCalibrationBody"
      responses:
        "201":
          description: Calibration started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdasCalibration"
...
  /programming/sessions:
    get:
      operationId: listProgrammingSessions
      tags: [programming]
      summary: List ECU programming sessions
      responses:
        "200":
          description: Programming sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgrammingSession"
    post:
      operationId: startProgrammingSession
      tags: [programming]
      summary: Start ECU programming or coding session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartProgrammingSessionBody"
      responses:
        "201":
          description: Programming session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammingSession"
  /maintenance/resets:
    get:
      operationId: listServiceResets
      tags: [maintenance]
      summary: List service reset records
      parameters:
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Service resets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceReset"
    post:
      operationId: performServiceReset
      tags: [maintenance]
      summary: Perform a service reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PerformServiceResetBody"
      responses:
        "201":
          description: Service reset performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceReset"
...
components:
  schemas:
...
    DiagnosticSession:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleName:
          type: string
        status:
          type: string
          enum: [running, completed, failed]
        dtcCount:
          type: integer
        systemsScanned:
          type: integer
        totalSystems:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        notes:
          type: string
      required:
        - id
        - vehicleId
        - status
        - startedAt
    StartDiagnosticSessionBody:
      type: object
      properties:
        vehicleId:
          type: integer
        scanType:
          type: string
          enum: [full_scan, quick_scan, system_scan]
        systems:
          type: array
          items:
            type: string
        notes:
          type: string
      required:
        - vehicleId
        - scanType
    DtcCode:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        code:
          type: string
        system:
          type: string
        severity:
          type: string
          enum: [critical, warning, info, pending]
        description:
          type: string
        possibleCauses:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, pending, cleared, historical]
        clearedAt:
          type: string
          format: date-time
        detectedAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - system
        - severity
        - description
        - status
        - detectedAt
...
    AdasCalibration:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleName:
          type: string
        system:
          type: string
        calibrationType:
          type: string
          enum: [static, dynamic, both]
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        result:
          type: string
          enum: [pass, fail, incomplete]
        notes:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - vehicleId
        - system
        - calibrationType
        - status
        - startedAt
    CreateAdasCalibrationBody:
      type: object
      properties:
        vehicleId:
          type: integer
        system:
          type: string
        calibrationType:
          type: string
          enum: [static, dynamic, both]
        notes:
          type: string
      required:
        - vehicleId
        - system
        - calibrationType
...
    ProgrammingSession:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleName:
          type: string
        type:
          type: string
          enum: [ecu_programming, ecu_coding, key_programming, immobilizer, adaptation]
        ecuModule:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        progressPercent:
          type: integer
        notes:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - vehicleId
        - type
        - ecuModule
        - status
        - startedAt
    StartProgrammingSessionBody:
      type: object
      properties:
        vehicleId:
          type: integer
        type:
          type: string
          enum: [ecu_programming, ecu_coding, key_programming, immobilizer, adaptation]
        ecuModule:
          type: string
        notes:
          type: string
      required:
        - vehicleId
        - type
        - ecuModule
    ServiceReset:
      type: object
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        vehicleName:
          type: string
        resetType:
          type: string
        status:
          type: string
          enum: [success, failed, pending]
        notes:
          type: string
        performedAt:
          type: string
          format: date-time
      required:
        - id
        - vehicleId
        - resetType
        - status
        - performedAt
    PerformServiceResetBody:
      type: object
      properties:
        vehicleId:
          type: integer
        resetType:
          type: string
        notes:
          type: string
      required:
        - vehicleId
        - resetType
...